@@ -9,5 +9,5 @@ public class HelloWorld {
 
     public static void main(String... args) {
         Logger logger=Logger.getLogger(HelloWorld.class.getName());
-        logger.info("This is a Hello World!");    }
+        logger.info("This is a module-using Hello World!");    }
 }
